#JC-954 Added ability for secured removal of branches: * Since Spring Security of our version does not allow creation of several <http> elements and thus configuring different authorization strategies for different URLs, had to implement this manually. In the future we'll update Spring Security to newer version and will try to implement this security logic with its means. * A method to search common users was added to dao & services because admin user is not registered within JCommune, but rather created in Poulpe. * Added logic to check the password in the PoulpeNotificationHandler
Tests were written only for dao & services, will need to write them for controller as well and also will implement security for removal of sections & comopnents.

+++ jcommune-view/jcommune-web-controller/src/main/java/org/jtalks/jcommune/web/controller/integration/PoulpeNotificationHandler.java
@@ -15,89 +15,103 @@
 
 package org.jtalks.jcommune.web.controller.integration;
 
+import org.jtalks.common.model.entity.User;
 import org.jtalks.jcommune.service.BranchService;
 import org.jtalks.jcommune.service.SectionService;
+import org.jtalks.jcommune.service.UserService;
 import org.jtalks.jcommune.service.exceptions.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
 import org.springframework.web.servlet.view.json.MappingJacksonJsonView;
 
+import static com.google.common.base.Preconditions.checkArgument;
+import static org.apache.commons.lang.StringUtils.isNotBlank;
+
 /**
- * Controller that handles notifications from Poulpe.
- * Main purpose of this is to support data consistency, e.g. when deleting
- * branch Poulpe can't take care of topics, proper post count, notifications and
- * so on. That is why messages are sent here.
+ * Controller that handles notifications from Poulpe. Main purpose of this is to support data consistency, e.g. when
+ * deleting branch Poulpe can't take care of topics, proper post count, notifications and so on. That is why messages
+ * are sent here.
  *
  * @author Vyacheslav Mishcheryakov
  * @author Evgeniy Naumenko
  */
 @Controller
 public class PoulpeNotificationHandler {
-    
     private static final String ERROR_MESSAGE_PARAMETER = "errorMessage";
-    
-    private BranchService branchService;
-    private SectionService sectionService;
-    
-    
+
+    private final BranchService branchService;
+    private final SectionService sectionService;
+    private final UserService userService;
+
     @Autowired
-    public PoulpeNotificationHandler(BranchService branchService,
-                            SectionService sectionService) {
+    public PoulpeNotificationHandler(BranchService branchService, SectionService sectionService, UserService userService) {
         this.branchService = branchService;
         this.sectionService = sectionService;
+        this.userService = userService;
     }
 
-
     /**
-     * Handles notification about branch deletion. This method deletes all 
-     * topics in branch. Branch itself is not deleted as Poulpe can cope with it
+     * Handles notification about branch deletion. This method deletes all topics in branch. Branch itself is not
+     * deleted as Poulpe can cope with it
      *
      * @param branchId branch id
      * @throws NotFoundException is thrown if branch not found
      */
-    @RequestMapping(value="/branches/{branchId}", method=RequestMethod.DELETE)
+    @RequestMapping(value = "/branches/{branchId}", method = RequestMethod.DELETE)
     @ResponseBody
-    public void deleteBranch(@PathVariable("branchId") long branchId) throws NotFoundException {
+    public void deleteBranch(@PathVariable("branchId") long branchId,
+                             @RequestParam(value = "password") String adminPassword) throws NotFoundException {
+        assertAdminPasswordCorrect(adminPassword);
         branchService.deleteAllTopics(branchId);
     }
-    
+
     /**
-     * Handles notification about section deletion. Removes all topics from
-     * section. Sections and branches won't be removed
+     *
+     * @param adminPassword
+     * @throws NotFoundException
+     */
+    private void assertAdminPasswordCorrect(String adminPassword) throws NotFoundException {
+        checkArgument(isNotBlank(adminPassword), "No password specified while it is required");
+        User admin = userService.getCommonUserByUsername("admin");
+        checkArgument(adminPassword.equals(admin.getPassword()),
+                "Wrong password was specified during removal of branch/section/component.");
+    }
+
+    /**
+     * Handles notification about section deletion. Removes all topics from section. Sections and branches won't be
+     * removed
+     *
      * @param sectionId section id
      * @throws NotFoundException is thrown if section not found
      */
-    @RequestMapping(value="/sections/{sectionId}", method=RequestMethod.DELETE)
+    @RequestMapping(value = "/sections/{sectionId}", method = RequestMethod.DELETE)
     @ResponseBody
     public void deleteSection(@PathVariable("sectionId") long sectionId) throws NotFoundException {
         sectionService.deleteAllTopicsInSection(sectionId);
     }
 
     /**
-     * Handles notification about component deletion.
-     * As for now it removes all the topics, leaving branches, sections and components untouched.
+     * Handles notification about component deletion. As for now it removes all the topics, leaving branches, sections
+     * and components untouched.
      *
-     * @throws org.jtalks.jcommune.service.exceptions.NotFoundException if object for deletion has not been found
+     * @throws org.jtalks.jcommune.service.exceptions.NotFoundException
+     *          if object for deletion has not been found
      */
-    @RequestMapping(value="/component", method=RequestMethod.DELETE)
+    @RequestMapping(value = "/component", method = RequestMethod.DELETE)
     @ResponseBody
     public void deleteComponent() throws NotFoundException {
         sectionService.deleteAllTopicsInForum();
     }
-    
+
     /**
-     * Catches all exceptions threw by any method in this controller and 
-     * returns HTTP status 500 with error message in response body instead.  
+     * Catches all exceptions threw by any method in this controller and returns HTTP status 500 with error message in
+     * response body instead.
+     *
      * @param exception exception that was thrown
-     * @return {@link ModelAndView} object with JSON view and error message as model 
+     * @return {@link ModelAndView} object with JSON view and error message as model
      */
     @ExceptionHandler(Exception.class)
     @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
